+++ src/components/App/App.jsx
@@ -2,13 +2,12 @@ import React, {useState, UseEffect, useEffect} from 'react';
 import axios from 'axios';
 import Header from '../Header/Header.jsx'
 import './App.css';
+import Form from '../Form/Form'
 
 
 function App() {
     let [groceryList, setGroceryList] = useState([]);
-    let [listItem, setListItem] = useState('');
-    let [itemQuantity, setItemQuantity] = useState('');
-    let [itemUnit, setItemUnit] = useState('');
+
 
     useEffect(() => {
         getList()
@@ -37,14 +36,7 @@ function App() {
         })
     }//end addItem
 
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        if (listItem && itemQuantity && itemUnit) {
-            addItem();
-        } else {
-            alert('Fill out the form pweeez')
-        }
-    }//end handle submit
+ 
 
     return (
         <div className="App">
@@ -52,16 +44,7 @@ function App() {
             <main>
                 <p>Under Construction...</p>
             </main>
-            <form onSubmit={handleSubmit}>
-                <h2>Add an Item!</h2>
-                <label>Item</label>
-                <input type="text" placeholder="Item" value={listItem} onChange={(event) => setListItem(event.target.value)} />
-                <label>Quantity</label>
-                <input type="number" placeholder="Quantity" value={itemQuantity} onChange={(event) => setItemQuantity(event.target.value)} />
-                <label>Unit</label>
-                <input type="text" placeholder="Unit" value={itemUnit} onChange={(event) => setItemUnit(event.target.value)} />
-                <button type="submit" value="Submit">Add Item</button>
-            </form>
+            <Form />
             <section>
                 <h2>Shopping List</h2>
                 <button>Reset</button>
